# public/chatgpt-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cultlike OS Executive Operating System API (ChatGPT & Claude Connectors)",
    "description": "Official API for ChatGPT Custom GPT Actions & Claude Connectors to orchestrate projects, sprint tasks, content vault deliverables, Fathom meetings, and execution analytics in Cultlike OS.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://callmy-mgmt.vercel.app",
      "description": "Cultlike OS Server"
    }
  ],
  "paths": {
...
    "/api/chatgpt/tasks": {
      "get": {
        "operationId": "listTasks",
        "summary": "List and search tasks with point-in-time date filtering",
        "description": "Lists tasks with optional filtering by status, priority, project_id, search keyword, or specific date (YYYY-MM-DD) for historical status.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "todo",
                "in_progress",
                "blocked",
                "shipped",
                "killed",
                "active"
              ]
            },
            "description": "Filter by task status."
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "p0",
                "p1",
                "p2",
                "p3"
              ]
            },
            "description": "Filter by priority level."
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks belonging to project UUID."
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search keyword matching title."
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Point-in-time historical date (YYYY-MM-DD)."
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks created after this ISO date."
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter tasks created before this ISO date."
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ]
            },
            "description": "Format of output: json or csv."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Max items."
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks."
          }
        }
      },
      "post": {
        "operationId": "createTask",
        "summary": "Create a new task",
        "description": "Creates a new task in CallMy Mgmt.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Task title."
                  },
                  "description": {
                    "type": "string",
                    "description": "Task description."
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "p0",
                      "p1",
                      "p2",
                      "p3"
                    ],
                    "default": "p2"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "todo",
                      "in_progress",
                      "blocked",
                      "shipped",
                      "killed"
                    ],
                    "default": "todo"
                  },
                  "project_id": {
                    "type": "string"
                  },
                  "due_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "time_box_minutes": {
                    "type": "integer",
                    "default": 60
                  },
                  "output_description": {
                    "type": "string"
                  },
                  "blocked_reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task created successfully."
          }
        }
      },
      "patch": {
        "operationId": "updateTask",
        "summary": "Update a task",
        "description": "Updates a task by task ID in the request body.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "Task UUID."
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "p0",
                      "p1",
                      "p2",
                      "p3"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "todo",
                      "in_progress",
                      "blocked",
                      "shipped",
                      "killed"
                    ]
                  },
                  "project_id": {
                    "type": "string"
                  },
                  "due_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "time_box_minutes": {
                    "type": "integer"
                  },
                  "output_description": {
                    "type": "string"
                  },
                  "blocked_reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}